# server/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: SkillHub API
  description: |
    SkillHub backend HTTP API.  Covers native portal routes (/api/v1/),
    web aliases (/api/web/), CLI routes (/api/cli/v1/), tool-facing
    routes for the miqro CLI (/api/tool/v1/), ClawHub compatibility,
    well-known discovery endpoints, and frontend page-oriented read
    models (/api/v1/frontend/*).
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/auth/login:
    post:
      summary: Local username/password login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Login successful; returns user principal. When SKILLHUB_SESSION_BACKEND=redis, a Set-Cookie header with an HttpOnly skillhub_session cookie is included.
          headers:
            Set-Cookie:
              description: Present when Redis-backed sessions are enabled. Contains an HttpOnly skillhub_session cookie.
              schema:
                type: string
        "401":
          description: Invalid credentials
...
  /api/v1/auth/register:
    post:
      summary: Register a new local user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
                displayName: { type: string }
                email: { type: string }
      responses:
        "201":
          description: User registered
        "409":
          description: Username or email already taken
  /api/v1/auth/me:
    get:
      summary: Returns the authenticated caller's identity
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Current user principal
        "401":
          description: Not authenticated
...
  /api/v1/namespaces/{slug}:
    get:
      summary: Get namespace by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Namespace detail

  # --- CLI ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API token (format: sk_...)"
    sessionCookie:
      type: apiKey
      in: cookie
      name: skillhub_session
...